+++ src/services/SpecStore.ts
@@ -28,7 +28,10 @@ export class SpecStore {
     this.externalDocs = this.parser.spec.externalDocs;
     this.contentItems = MenuBuilder.buildStructure(this.parser, this.options);
     this.securitySchemes = new SecuritySchemesModel(this.parser);
-    const webhookPath: Referenced<OpenAPIPath> = {...this.parser?.spec?.['x-webhooks'], ...this.parser?.spec.webhooks};
+    const webhookPath: Referenced<OpenAPIPath> = {
+      ...this.parser?.spec?.['x-webhooks'],
+      ...this.parser?.spec.webhooks,
+    };
     this.webhooks = new WebhookModel(this.parser, options, webhookPath);
   }
 }