+++ src/services/models/Schema.ts
@@ -134,7 +134,10 @@ export class SchemaModel {
     this.maxItems = schema.maxItems;
 
     if (!!schema.nullable || schema['x-nullable']) {
-      if (Array.isArray(this.type) && !this.type.some((value) => value === null || value === 'null')) {
+      if (
+        Array.isArray(this.type) &&
+        !this.type.some(value => value === null || value === 'null')
+      ) {
         this.type = [...this.type, 'null'];
       } else if (!Array.isArray(this.type) && (this.type !== null || this.type !== 'null')) {
         this.type = [this.type, 'null'];
@@ -142,7 +145,7 @@ export class SchemaModel {
     }
 
     this.displayType = Array.isArray(this.type)
-      ? this.type.map(item => item === null ? 'null' : item).join(' or ')
+      ? this.type.map(item => (item === null ? 'null' : item)).join(' or ')
       : this.type;
 
     if (this.isCircular) {
@@ -155,7 +158,7 @@ export class SchemaModel {
     } else if (
       isChild &&
       Array.isArray(schema.oneOf) &&
-      schema.oneOf.find((s) => s.$ref === this.pointer)
+      schema.oneOf.find(s => s.$ref === this.pointer)
     ) {
       // we hit allOf of the schema with the parent discriminator
       delete schema.oneOf;
@@ -195,8 +198,7 @@ export class SchemaModel {
       }
       if (Array.isArray(this.type)) {
         const filteredType = this.type.filter(item => item !== 'array');
-        if (filteredType.length)
-          this.displayType += ` or ${filteredType.join(' or ')}`;
+        if (filteredType.length) this.displayType += ` or ${filteredType.join(' or ')}`;
       }
     }
 
@@ -215,7 +217,7 @@ export class SchemaModel {
       const title =
         isNamedDefinition(variant.$ref) && !merged.title
           ? JsonPointer.baseName(variant.$ref)
-          : `${(merged.title || '')}${(merged.const && JSON.stringify(merged.const)) || ''}`;
+          : `${merged.title || ''}${(merged.const && JSON.stringify(merged.const)) || ''}`;
 
       const schema = new SchemaModel(
         parser,
@@ -243,7 +245,7 @@ export class SchemaModel {
       this.displayType = types.join(' or ');
     } else {
       this.displayType = this.oneOf
-        .map((schema) => {
+        .map(schema => {
           let name =
             schema.typePrefix +
             (schema.title ? `${schema.title} (${schema.displayType})` : schema.displayType);
@@ -364,7 +366,7 @@ function buildFields(
   const props = schema.properties || {};
   const additionalProps = schema.additionalProperties;
   const defaults = schema.default;
-  let fields = Object.keys(props || []).map((fieldName) => {
+  let fields = Object.keys(props || []).map(fieldName => {
     let field = props[fieldName];
 
     if (!field) {