+++ src/utils/index.ts
@@ -8,3 +8,4 @@ export * from './dom';
 export * from './decorators';
 export * from './debug';
 export * from './memoize';
+export * from './sort';